Make mutt-nntp depend on mutt >= 1.5.21-5
Since Ubuntu uses a revision -5ubuntu, and our package revision,
-5nntpX, revision is lexicographically lower than that, using a version
specification of <= fails to install mutt from Ubuntu with mutt-nntp.
Using a specification of >= will fail on Debian (which uses just -5).

Instead we use a manually specified revision matching the Debian
package, as is the one only with a numeric revision which is also used
by Ubuntu. Now mutt-nntp requires mutt >= 1.5.21-5, which should be
satisfied by both Debian and Ubuntu (1.5.21-5ubuntuX).